requirements.txt
curl_cffi>=0.7
python-telegram-bot[webhooks]>=21.0
beautifulsoup4>=4.12
lxml>=5.0
PyYAML>=6.0
